requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
aws-requests-auth==0.4.3
beautifulsoup4==4.14.3
boto3==1.42.0
botocore==1.41.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dill==0.4.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.123.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
frozenlist==1.8.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
log-symbols==0.0.14
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
mcp==1.22.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
ollama==0.6.1
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.30.0
s3transfer==0.16.0
sentry-sdk==2.46.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.27.0
slack_sdk==3.39.0
soupsieve==2.8
spinners==0.0.24
sse-starlette==3.0.3
starlette==0.50.0
strands-agents==1.18.0
strands-agents-builder==0.1.10
strands-agents-tools==0.2.16
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
